refactor: @DritiesContext를 @Transactional로 수정
